ConsecutiveBounceHandler
