+++ src/Domain/Messaging/Command/ProcessBouncesCommand.php
@@ -5,17 +5,11 @@ declare(strict_types=1);
 namespace PhpList\Core\Domain\Messaging\Command;
 
 use Exception;
-use PhpList\Core\Domain\Messaging\Model\Bounce;
-use PhpList\Core\Domain\Messaging\Model\UserMessage;
-use PhpList\Core\Domain\Messaging\Model\UserMessageBounce;
+use PhpList\Core\Domain\Messaging\Service\ConsecutiveBounceHandler;
 use PhpList\Core\Domain\Messaging\Service\Processor\BounceProtocolProcessor;
 use PhpList\Core\Domain\Messaging\Service\Processor\AdvancedBounceRulesProcessor;
 use PhpList\Core\Domain\Messaging\Service\Processor\UnidentifiedBounceReprocessor;
 use PhpList\Core\Domain\Messaging\Service\LockService;
-use PhpList\Core\Domain\Messaging\Service\Manager\BounceManager;
-use PhpList\Core\Domain\Subscription\Repository\SubscriberRepository;
-use PhpList\Core\Domain\Subscription\Service\Manager\SubscriberHistoryManager;
-use PhpList\Core\Domain\Subscription\Service\Manager\SubscriberManager;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
@@ -47,16 +41,13 @@ class ProcessBouncesCommand extends Command
     }
 
     public function __construct(
-        private readonly BounceManager $bounceManager,
         private readonly LockService $lockService,
         private readonly LoggerInterface $logger,
-        private readonly SubscriberManager $subscriberManager,
-        private readonly SubscriberHistoryManager $subscriberHistoryManager,
-        private readonly SubscriberRepository $subscriberRepository,
         /** @var iterable<BounceProtocolProcessor> */
         private readonly iterable $protocolProcessors,
         private readonly AdvancedBounceRulesProcessor $advancedRulesProcessor,
         private readonly UnidentifiedBounceReprocessor $unidentifiedBounceReprocessor,
+        private readonly ConsecutiveBounceHandler $consecutiveBounceHandler,
     ) {
         parent::__construct();
     }
@@ -83,6 +74,8 @@ class ProcessBouncesCommand extends Command
         try {
             $io->title('Processing bounces');
             $protocol = (string)$input->getOption('protocol');
+            $unsubscribeThreshold = (int)$input->getOption('unsubscribe-threshold');
+            $blacklistThreshold = (int)$input->getOption('blacklist-threshold');
 
             $downloadReport = '';
 
@@ -103,15 +96,9 @@ class ProcessBouncesCommand extends Command
             $downloadReport .= $processor->process($input, $io);
             $this->unidentifiedBounceReprocessor->process($io);
             $this->advancedRulesProcessor->process($io, (int)$input->getOption('rules-batch-size'));
-
-            $this->handleConsecutiveBounces(
-                $io,
-                (int)$input->getOption('unsubscribe-threshold'),
-                (int)$input->getOption('blacklist-threshold')
-            );
+            $this->consecutiveBounceHandler->handle($io, $unsubscribeThreshold, $blacklistThreshold);
 
             $this->logger->info('Bounce processing completed', ['downloadReport' => $downloadReport]);
-
             $io->success('Bounce processing completed.');
 
             return Command::SUCCESS;
@@ -123,60 +110,5 @@ class ProcessBouncesCommand extends Command
         } finally {
             $this->lockService->release($lock);
         }
-    }
-
-    private function handleConsecutiveBounces(SymfonyStyle $io, int $unsubscribeThreshold, int $blacklistThreshold): void
-    {
-        $io->section('Identifying consecutive bounces');
-        $users = $this->subscriberRepository->distinctUsersWithBouncesConfirmedNotBlacklisted();
-        $total = count($users);
-        if ($total === 0) {
-            $io->writeln('Nothing to do');
-            return;
-        }
-        $usercnt = 0;
-        foreach ($users as $user) {
-            $usercnt++;
-            $history = $this->bounceManager->getUserMessageHistoryWithBounces($user);
-            $cnt = 0; $removed = false; $msgokay = false; $unsubscribed = false;
-            foreach ($history as $bounce) {
-                /** @var $bounce array{um: UserMessage, umb: UserMessageBounce|null, b: Bounce|null} */
-                if (
-                    stripos($bounce['b']->getStatus() ?? '', 'duplicate') === false
-                    && stripos($bounce['b']->getComment() ?? '', 'duplicate') === false
-                ) {
-                    if ($bounce['b']->getId()) {
-                        $cnt++;
-                        if ($cnt >= $unsubscribeThreshold) {
-                            if (!$unsubscribed) {
-                                $this->subscriberManager->markUnconfirmed($user->getId());
-                                $this->subscriberHistoryManager->addHistory(
-                                    subscriber: $user,
-                                    message: 'Auto Unconfirmed',
-                                    details: sprintf('Subscriber auto unconfirmed for %d consecutive bounces', $cnt)
-                                );
-                                $unsubscribed = true;
-                            }
-                            if ($blacklistThreshold > 0 && $cnt >= $blacklistThreshold) {
-                                $this->subscriberManager->blacklist(
-                                    subscriber: $user,
-                                    reason: sprintf('%d consecutive bounces, threshold reached', $cnt)
-                                );
-                                $removed = true;
-                            }
-                        }
-                    } else {
-                        break;
-                    }
-                }
-                if ($removed) {
-                    break;
-                }
-            }
-            if ($usercnt % 5 === 0) {
-                $io->writeln(sprintf('processed %d out of %d subscribers', $usercnt, $total));
-            }
-        }
-        $io->writeln(sprintf('total of %d subscribers processed', $total));
     }
 }
